Do double-dot-expand only in certain conditions
We want to expand at the beginning of the line, after a space, and after
a `/`.
Also, we want to expand after `|` (pipeline), `&&` and `||` (separators
in a sublist), `;` `&` `&|` and `&!` (terminators of a sublist), `!`
(history expansion), `<` `>` `>|` `>!` and etc. (redirection). This can
be simplified to expanding after any of: `!` `&` `/` `;` `<` `>` `|`

I'm probably missing some other scenarios where we consider there can be
a file as input...  :- /

Fixes #3